[microNPU] Upgrade Vela to v3.2.0 (#9635)
* [microNPU] Upgrade Vela to v3.2.0

In addition to upgrading the version of Vela, this contains a bug fix
and a fix due to an api change.

The bug fix was found as a result of upgrading the version and ensures
that block traversal mode is calculated before a set of block configs
is found. As a result, valid block configs are calculated consistently
throughout compilation.

The fix for the api change is a cast to the datatype now expected by
Vela due to using strict casting rules.


* adding version change to gen requirements

+++ python/tvm/relay/backend/contrib/ethosu/vela_api.py
@@ -207,7 +207,7 @@ def compress_weights(
     layout_transform_indices = {"HWIO": (3, 0, 1, 2), "HWOI": (2, 0, 1, 3), "OHWI": (0, 1, 2, 3)}
     assert weights_layout in layout_transform_indices.keys()
     assert isinstance(weights_zp, np.int64)
-    weights = weights.astype(np.int64) - weights_zp
+    weights = weights.astype(np.int16) - weights_zp
     # Vela needs the weights in OHWI layout
     weights_ohwi = np.transpose(weights, layout_transform_indices[weights_layout])
     shape_ohwi = [
